step 04: separate classes
